% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smcDE.R
\name{smcDE}
\alias{smcDE}
\title{This function estimate parameters in DEs via SMC based on collocation}
\usage{
smcDE(
  data,
  times,
  seed,
  knots,
  CESSthresholds,
  NP,
  resampleThreshold,
  alambda,
  blambda,
  sigmac,
  DEmodel
)
}
\arguments{
\item{data}{a list with each element being a trajectory with measurement error}

\item{times}{a list with each element being the time we record data}

\item{seed}{the random seeds for propagation and resampling in SMC}

\item{CESSthresholds}{relative conditional effective sample size (0, 1) in SMC}

\item{NP}{the number of particles in SMC}

\item{resampleThreshold}{the threshold triggering resampling in SMC}

\item{alambda}{hyper-parameter for tuning parameter lambda}

\item{blambda}{hyper-parameter for tuning parameter lambda}

\item{sigmac}{the standard deviation of the basis coefficients c}

\item{DEmodel}{options for DE models: 1 (ODE), 2 (DDE1), 3 (DDE2) in manuscript}

\item{knotsPosition}{the location we put knots}
}
\description{
This function estimate parameters in DEs via SMC based on collocation
}
